requirements.txt
ola >= 0.10.7
RPi.GPIO
astral
tzlocal
pytest
schedule
flask